# Unification of bitmap unserialization by using auxiliary byteInput approach