--- a/byte_input.go
+++ b/byte_input.go
@@ -3,64 +3,48 @@ package roaring
 import (
 	"encoding/binary"
 	"io"
-	"io/ioutil"
 )
-
-type byteInput struct {
-	r         io.Reader
-	readBytes int
-}
 
 type byteBuffer interface {
 	// Next returns a slice containing the next n bytes from the buffer,
 	// advancing the buffer as if the bytes had been returned by Read.
-	Next(n int) []byte
+	next(n int) ([]byte, error)
 }
 
-// Read reads up to len(p) bytes into p. It returns the number of bytes
-// read (0 <= n <= len(p)) and any error encountered.
-func (b *byteInput) Read(p []byte) (n int, err error) {
-	return b.r.Read(p)
+type byteInput struct {
+	buf       byteBuffer
+	readBytes int
+}
+
+func newByteInputFromReader(reader io.Reader) *byteInput {
+	return &byteInput{
+		buf:       &byteBufferAdapter{reader},
+		readBytes: 0,
+	}
+}
+
+func newByteInputFromBuffer(buf []byte) *byteInput {
+	return &byteInput{
+		buf:       &byteBufferImpl{buf: buf, off: 0},
+		readBytes: 0,
+	}
 }
 
 // Next returns a slice containing the next n bytes from the reader
 // If there are fewer bytes than the given n, io.ErrUnexpectedEOF will be returned
 func (b *byteInput) next(n int) ([]byte, error) {
-	if buf, ok := b.r.(byteBuffer); ok {
-		data := buf.Next(n)
-		b.readBytes += len(data)
-
-		if len(data) != n {
-			return nil, io.ErrUnexpectedEOF
-		}
-
-		return data, nil
-	}
-
-	buf := make([]byte, n)
-	readBytes, err := b.r.Read(buf)
-	b.readBytes += readBytes
+	data, err := b.buf.next(n)
+	b.readBytes += len(data)
 
 	if err != nil {
 		return nil, err
 	}
 
-	if readBytes != n {
+	if len(data) != n {
 		return nil, io.ErrUnexpectedEOF
 	}
 
-	return buf[:n], nil
-}
-
-// ReadUInt64 reads uint64 with LittleEndian order
-func (b *byteInput) readUInt64() (uint64, error) {
-	buf, err := b.next(8)
-
-	if err != nil {
-		return 0, err
-	}
-
-	return binary.LittleEndian.Uint64(buf), nil
+	return data, nil
 }
 
 // ReadUInt64 reads uint32 with LittleEndian order
@@ -92,23 +76,53 @@ func (b *byteInput) getReadBytes() int64 {
 
 // skipBytes skips exactly n bytes, if
 func (b *byteInput) skipBytes(n int) (err error) {
-	if buf, ok := b.r.(byteBuffer); ok {
-		data := buf.Next(n)
-		b.readBytes += len(data)
+	data, err := b.buf.next(n)
+	b.readBytes += len(data)
 
-		if len(data) != n {
-			err = io.ErrUnexpectedEOF
-		}
-
+	if err != nil {
 		return
 	}
 
-	readBytes, err := io.CopyN(ioutil.Discard, b.r, int64(n))
-	b.readBytes += int(readBytes)
-
-	if int(readBytes) != n {
+	if len(data) != n {
 		err = io.ErrUnexpectedEOF
 	}
 
 	return
+}
+
+type byteBufferAdapter struct {
+	r io.Reader
+}
+
+// Next returns a slice containing the next n bytes from the buffer,
+// advancing the buffer as if the bytes had been returned by Read.
+func (b *byteBufferAdapter) next(n int) ([]byte, error) {
+	buf := make([]byte, n)
+	readBytes, err := b.r.Read(buf)
+
+	if err != nil {
+		return nil, err
+	}
+
+	return buf[:readBytes], nil
+}
+
+type byteBufferImpl struct {
+	buf []byte
+	off int
+}
+
+// Next returns a slice containing the next n bytes from the buffer,
+// advancing the buffer as if the bytes had been returned by Read.
+func (b *byteBufferImpl) next(n int) ([]byte, error) {
+	m := len(b.buf) - b.off
+
+	if n > m {
+		n = m
+	}
+
+	data := b.buf[b.off : b.off+n]
+	b.off += n
+
+	return data, nil
 }