… byte_input_test.go
… serialization.go
… serialization_generic.go
… serialization_littleendian.go
… serialization_test.go
… testdata/crashwithinvalidcookie.bin
… unserialize_test.go
… util.go