This pull request is not aimed to be merged, it should be treated as a proposal for enhancement.

Currently, bitmap package has two cases that I would like to consider:

* Methods `readFrom/fromBuffer` are responsible for restoring bitmap's state from `io.Reader` and `byte slice`. They have pretty the same code base and each of them has own cons. 

* `fromBuffer` makes a user keep (to hold reference) the provided byte slice and also to leave it constant. 

* `readFrom` loses the performance due to the necessity to create objects from the read byte data.

This pull request tries to solve these issues. For unification was implemented the auxiliary class `byteInput` for io utils. Also, in `roaringArray` was added special field `sentry` that holds the references of `allocated/type converted` slices (here I am not sure about this approach, in theory, it should work and all we need to ask a user to keep the provided `buf` immutable).