@lemire 2019-08-07
The goal is reduce the logic duplication. Do I understand correctly?