@alldroll 2019-08-07
Yes, also possibly save a user from the need to hold a reference on `buf` byte slice for method `FromBuffer`