@lemire 2019-08-07
This looks like an interesting design!

1. How do you handle big endian systems. The Roaring format is little endian, and bytes must be flipped on big endian systems...

2. You have deleted some tests without adding new ones... ideally we want more testing over time, not less...

3. You claim that this change would free us from having to hold on to the slice with FromBuffer. Can you elaborate? I You either have to make a copy or hold on to a reference.

4. Your code is more elegant and maintainable. This is important but we don’t want to sacrifice performance. Have you checked the performance? Before merging this, we would want to make sure that there is no performance regression.