@alldroll 2019-08-08
I totally agree with you, I have no questions about 2,3,4 cases. But could you describe more the case with `big endian systems` . How was it handled before? All I have done is just leave only `FromBuffer` implementation with `bytesBuffer` instead of `byte slice`.