@lemire 2019-08-08
Look at the file diff ... we had calls to binary.LittleEndian.Uint16 on big endian systems. This may be gone.

I think we don’t test on big endian systems in the CI.