@lemire 2019-08-08
We need to find a way to test on big endian.

How is the performance in your tests!