@lemire 2019-08-08
Ok. I have done some reading and it appears that Go does not yet support big endian systems (which are uncommon anyhow) so let us forget about the endianness question.