@alldroll 2019-08-09
I wrote some benchmarks, here it is

```
benchmark                                               old ns/op     new ns/op     delta
BenchmarkUnserializeReadFrom/ReadFrom-650-8             1367          1031          -24.58%
BenchmarkUnserializeReadFrom/ReadFrom-6500-8            5024          3014          -40.01%
BenchmarkUnserializeReadFrom/ReadFrom-65000-8           5175          2939          -43.21%
BenchmarkUnserializeReadFrom/ReadFrom-650000-8          14214         5055          -64.44%
BenchmarkUnserializeReadFrom/ReadFrom-6500000-8         66036         19665         -70.22%
BenchmarkUnserializeFromBuffer/FromBuffer-650-8         166           236           +42.17%
BenchmarkUnserializeFromBuffer/FromBuffer-6500-8        157           263           +67.52%
BenchmarkUnserializeFromBuffer/FromBuffer-65000-8       166           280           +68.67%
BenchmarkUnserializeFromBuffer/FromBuffer-650000-8      1034          1378          +33.27%
BenchmarkUnserializeFromBuffer/FromBuffer-6500000-8     8363          11271         +34.77%

benchmark                                               old allocs     new allocs     delta
BenchmarkUnserializeReadFrom/ReadFrom-650-8             14             11             -21.43%
BenchmarkUnserializeReadFrom/ReadFrom-6500-8            14             11             -21.43%
BenchmarkUnserializeReadFrom/ReadFrom-65000-8           14             11             -21.43%
BenchmarkUnserializeReadFrom/ReadFrom-650000-8          94             38             -59.57%
BenchmarkUnserializeReadFrom/ReadFrom-6500000-8         823            308            -62.58%
BenchmarkUnserializeFromBuffer/FromBuffer-650-8         4              5              +25.00%
BenchmarkUnserializeFromBuffer/FromBuffer-6500-8        4              5              +25.00%
BenchmarkUnserializeFromBuffer/FromBuffer-65000-8       4              5              +25.00%
BenchmarkUnserializeFromBuffer/FromBuffer-650000-8      13             14             +7.69%
BenchmarkUnserializeFromBuffer/FromBuffer-6500000-8     103            104            +0.97%

benchmark                                               old bytes     new bytes     delta
BenchmarkUnserializeReadFrom/ReadFrom-650-8             1614          1540          -4.58%
BenchmarkUnserializeReadFrom/ReadFrom-6500-8            8423          8324          -1.18%
BenchmarkUnserializeReadFrom/ReadFrom-65000-8           8422          8324          -1.16%
BenchmarkUnserializeReadFrom/ReadFrom-650000-8          18223         9096          -50.09%
BenchmarkUnserializeReadFrom/ReadFrom-6500000-8         32625         16794         -48.52%
BenchmarkUnserializeFromBuffer/FromBuffer-650-8         52            84            +61.54%
BenchmarkUnserializeFromBuffer/FromBuffer-6500-8        52            84            +61.54%
BenchmarkUnserializeFromBuffer/FromBuffer-65000-8       52            84            +61.54%
BenchmarkUnserializeFromBuffer/FromBuffer-650000-8      672           704           +4.76%
BenchmarkUnserializeFromBuffer/FromBuffer-6500000-8     6896          6928          +0.46%
```
I can't say, that this pull request really does а good job, as we could see, we have some performance benefits for `ReadFrom` method, but also have a slight reduction for `FromBuffer`

Also, I do have concern for the method `byteBufferAdapter.next()`. Here we allocate a new memory block for each call and use `byteSliceAsUint16Slice` instead of creating new objects [as here](https://github.com/alldroll/roaring/blob/master/roaringarray.go#L721). I can't find any proof that this will work properly and Go will keep a reference on the allocated memory block. I tried to catch it in tests `TestHoldReference`, but it seems to me that I do something wrong because of misunderstanding how Go works with memory.