@lemire 2019-08-09
Thanks for the thorough benchmarks. I particularly appreciate the various metrics you have used.

I think you have shown that your PR has potential value and should be pursued as you improve very significantly the performance. Of course, we may ask whether it is safe...

Can't we test what happens with the garbage collector with `runtime.GC()`. It may not be perfect, but I think it could help stress your test.

You have a small performance with FromBuffer. Any chance we can narrow down the source of this performance regression? I think it should be possible, in theory at least, to have no performance degradation.

I think it would be acceptable to see a small (say 5%) performance regression given that we cut down the number of lines and thus improve maintainability. We probably want to avoid a performance degradation of more than 10%...

Please keep going... this looks very promising!