@lemire 2019-08-09
My gut feeling is that methods like `readUInt32()` might be where the performance degradation comes from.