@alldroll 2019-08-10
You are right. I have optimized `readUInt32`, `readUInt16`, `next` methods, and we have some achievement performance, but it is not sufficient yet
```
benchmark                                               old ns/op     new ns/op     delta
BenchmarkUnserializeReadFrom/ReadFrom-650-8             822           614           -25.30%
BenchmarkUnserializeReadFrom/ReadFrom-6500-8            2823          1734          -38.58%
BenchmarkUnserializeReadFrom/ReadFrom-65000-8           2856          1753          -38.62%
BenchmarkUnserializeReadFrom/ReadFrom-650000-8          8338          3044          -63.49%
BenchmarkUnserializeReadFrom/ReadFrom-6500000-8         43401         14037         -67.66%
BenchmarkUnserializeFromBuffer/FromBuffer-650-8         125           178           +42.40%
BenchmarkUnserializeFromBuffer/FromBuffer-6500-8        121           176           +45.45%
BenchmarkUnserializeFromBuffer/FromBuffer-65000-8       121           176           +45.45%
BenchmarkUnserializeFromBuffer/FromBuffer-650000-8      643           777           +20.84%
BenchmarkUnserializeFromBuffer/FromBuffer-6500000-8     5378          6180          +14.91%
```

I carry on working on it...