@lemire 2019-08-10
Thanks for adding runtime.GC, this gives us extra confidence.