@alldroll 2019-08-13
I've done some investigation and found that the cause of performance regression is an allocation of a new object byteInput. Unfortunately, I have no idea how to make it faster, there is only one thing came to my mind - using sync.Pool. This approach makes me concerned because of the code became more complex and confusing. Here are new benchmarks

```
benchmark                                               old ns/op     new ns/op     delta
BenchmarkUnserializeReadFrom/ReadFrom-650-8             1367          692           -49.38%
BenchmarkUnserializeReadFrom/ReadFrom-6500-8            5024          2428          -51.67%
BenchmarkUnserializeReadFrom/ReadFrom-65000-8           5175          2356          -54.47%
BenchmarkUnserializeReadFrom/ReadFrom-650000-8          14214         4886          -65.63%
BenchmarkUnserializeReadFrom/ReadFrom-6500000-8         66036         22769         -65.52%
BenchmarkUnserializeFromBuffer/FromBuffer-650-8         166           214           +28.92%
BenchmarkUnserializeFromBuffer/FromBuffer-6500-8        157           212           +35.03%
BenchmarkUnserializeFromBuffer/FromBuffer-65000-8       166           206           +24.10%
BenchmarkUnserializeFromBuffer/FromBuffer-650000-8      1034          1071          +3.58%
BenchmarkUnserializeFromBuffer/FromBuffer-6500000-8     8363          9360          +11.92%
```