@alldroll 2019-08-21
@lemire 

I've run the same benchmarks with a flag `-tags appengine` and here is what we got:

```
benchmark                                               old ns/op     new ns/op     delta
BenchmarkUnserializeReadFrom/ReadFrom-650-8             5503          2804          -49.05%
BenchmarkUnserializeReadFrom/ReadFrom-6500-8            13344         7364          -44.81%
BenchmarkUnserializeReadFrom/ReadFrom-65000-8           13721         6750          -50.81%
BenchmarkUnserializeReadFrom/ReadFrom-650000-8          22866         9661          -57.75%
BenchmarkUnserializeReadFrom/ReadFrom-6500000-8         75620         28100         -62.84%
BenchmarkUnserializeFromBuffer/FromBuffer-650-8         1984          1978          -0.30%
BenchmarkUnserializeFromBuffer/FromBuffer-6500-8        4231          4834          +14.25%
BenchmarkUnserializeFromBuffer/FromBuffer-65000-8       4299          4382          +1.93%
BenchmarkUnserializeFromBuffer/FromBuffer-650000-8      5892          6694          +13.61%
BenchmarkUnserializeFromBuffer/FromBuffer-6500000-8     16862         17084         +1.32%

benchmark                                               old allocs     new allocs     delta
BenchmarkUnserializeReadFrom/ReadFrom-650-8             16             12             -25.00%
BenchmarkUnserializeReadFrom/ReadFrom-6500-8            16             12             -25.00%
BenchmarkUnserializeReadFrom/ReadFrom-65000-8           16             12             -25.00%
BenchmarkUnserializeReadFrom/ReadFrom-650000-8          96             48             -50.00%
BenchmarkUnserializeReadFrom/ReadFrom-6500000-8         825            408            -50.55%
BenchmarkUnserializeFromBuffer/FromBuffer-650-8         6              6              +0.00%
BenchmarkUnserializeFromBuffer/FromBuffer-6500-8        6              6              +0.00%
BenchmarkUnserializeFromBuffer/FromBuffer-65000-8       6              6              +0.00%
BenchmarkUnserializeFromBuffer/FromBuffer-650000-8      24             24             +0.00%
BenchmarkUnserializeFromBuffer/FromBuffer-6500000-8     204            204            +0.00%

benchmark                                               old bytes     new bytes     delta
BenchmarkUnserializeReadFrom/ReadFrom-650-8             3059          2936          -4.02%
BenchmarkUnserializeReadFrom/ReadFrom-6500-8            16672         16508         -0.98%
BenchmarkUnserializeReadFrom/ReadFrom-65000-8           16671         16508         -0.98%
BenchmarkUnserializeReadFrom/ReadFrom-650000-8          26476         17364         -34.42%
BenchmarkUnserializeReadFrom/ReadFrom-6500000-8         40868         25791         -36.89%
BenchmarkUnserializeFromBuffer/FromBuffer-650-8         1464          1464          +0.00%
BenchmarkUnserializeFromBuffer/FromBuffer-6500-8        8248          8250          +0.02%
BenchmarkUnserializeFromBuffer/FromBuffer-65000-8       8248          8250          +0.02%
BenchmarkUnserializeFromBuffer/FromBuffer-650000-8      8944          8946          +0.02%
BenchmarkUnserializeFromBuffer/FromBuffer-6500000-8     15900         15904         +0.03%

```

There will no be performance reduction for the `appengine`